# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "dPO API",
    "version": "1.0.0",
    "description": "dPO API with autogenerated swagger doc"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/profile/{userId}/overview/header": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve the profile of the user",
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the user."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/api/v1/profile/{userId}/overview/info": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve the profile info of the user",
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the user."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/api/v1/profile/{userId}/renter-profile/": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retrieve the profile info of the renter",
        "tags": [
          "Profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the renter."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
    "/api/v1/support/": {
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "Get all support questions.",
        "responses": {
          "200": {
            "description": "A list of support questions."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Support"
        ],
        "summary": "Add a new support question.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "tag": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Support question added successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/v1/support/tag/{tag}": {
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "Get support questions by tag.",
        "parameters": [
          {
            "in": "path",
            "name": "tag",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Tag of the support question (e.g., renters, partners, others)."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of support questions."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/v1/support/search": {
      "get": {
        "tags": [
          "Support"
        ],
        "summary": "Search support questions by title.",
        "parameters": [
          {
            "in": "query",
            "name": "title",
            "schema": {
              "type": "string"
            },
            "description": "Keyword to search in the title."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matched support questions."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/api/v1/support/{id}": {
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "Support"
        ],
        "summary": "Delete a support question by ID.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID of the support question to delete."
          }
        ],
        "responses": {
          "204": {
            "description": "Support question deleted successfully."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
...
}